@@ -1,14 +1,35 @@
 module github.com/llaoj/oauth2nsso
 
-go 1.13
+go 1.24.3
 
 require (
-	github.com/go-ldap/ldap/v3 v3.4.2
-	github.com/go-oauth2/oauth2/v4 v4.4.2
+	github.com/go-ldap/ldap/v3 v3.4.11
+	github.com/go-oauth2/oauth2/v4 v4.5.3
 	github.com/golang-jwt/jwt v3.2.2+incompatible
-	github.com/gorilla/sessions v1.2.0
-	github.com/tidwall/gjson v1.9.3 // indirect
-	gopkg.in/yaml.v2 v2.3.0
-	gorm.io/driver/mysql v1.3.2
-	gorm.io/gorm v1.23.1
+	github.com/gorilla/sessions v1.4.0
+	gopkg.in/yaml.v2 v2.4.0
+	gorm.io/driver/mysql v1.6.0
+	gorm.io/gorm v1.30.0
+)
+
+require (
+	filippo.io/edwards25519 v1.1.0 // indirect
+	github.com/Azure/go-ntlmssp v0.0.0-20221128193559-754e69321358 // indirect
+	github.com/go-asn1-ber/asn1-ber v1.5.8-0.20250403174932-29230038a667 // indirect
+	github.com/go-sql-driver/mysql v1.9.3 // indirect
+	github.com/golang-jwt/jwt/v5 v5.2.2 // indirect
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/gorilla/securecookie v1.1.2 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+	github.com/tidwall/btree v0.0.0-20191029221954-400434d76274 // indirect
+	github.com/tidwall/buntdb v1.1.2 // indirect
+	github.com/tidwall/gjson v1.12.1 // indirect
+	github.com/tidwall/grect v0.0.0-20161006141115-ba9a043346eb // indirect
+	github.com/tidwall/match v1.1.1 // indirect
+	github.com/tidwall/pretty v1.2.0 // indirect
+	github.com/tidwall/rtree v0.0.0-20180113144539-6cd427091e0e // indirect
+	github.com/tidwall/tinyqueue v0.0.0-20180302190814-1e39f5511563 // indirect
+	golang.org/x/crypto v0.39.0 // indirect
+	golang.org/x/text v0.26.0 // indirect
 )
