@@ -1,42 +1,39 @@
 package config
 
 import (
-    // "fmt"
-    "strings"
+	"strings"
 )
 
 func Get() *App {
-    return &cfg
+	return &cfg
 }
 
 func GetOAuth2Client(clientID string) (cli OAuth2Client) {
-    for _, v := range cfg.OAuth2.Client {
-        if v.ID == clientID {
-            cli = v
-        }
-    }
-
-    return
+	for _, v := range cfg.OAuth2.Client {
+		if v.ID == clientID {
+			cli = v
+		}
+	}
+	return
 }
 
 func ScopeJoin(scope []Scope) string {
-    var s []string
-    for _, sc := range scope {
-        s = append(s, sc.ID)
-    }
-    return strings.Join(s, ",")
+	var s []string
+	for _, sc := range scope {
+		s = append(s, sc.ID)
+	}
+	return strings.Join(s, ",")
 }
 
 func ScopeFilter(clientID string, scope string) (s []Scope) {
-    cli := GetOAuth2Client(clientID)
-    sl := strings.Split(scope, ",")
-    for _, str := range sl {
-        for _, sc := range cli.Scope {
-            if str == sc.ID {
-                s = append(s, sc)
-            }
-        }
-    }
-
-    return
+	cli := GetOAuth2Client(clientID)
+	sl := strings.Split(scope, ",")
+	for _, str := range sl {
+		for _, sc := range cli.Scope {
+			if str == sc.ID {
+				s = append(s, sc)
+			}
+		}
+	}
+	return
 }
